Implement Department, update seeder and user/role permission

+++ resources/views/roles/index.blade.php
@@ -8,7 +8,9 @@
             <h2>Role Management</h2>
         </div>
         <div class="pull-right">
+            @can('role-create')
             <a class="btn btn-success" href="{{ route('roles.create') }}"> Create New Role</a>
+            @endcan
         </div>
     </div>
 </div>
@@ -34,11 +36,15 @@
         <td>
             <a class="btn btn-info" href="{{ route('roles.show',$role->id) }}">Show</a>
             
+            @can('role-edit')
                 <a class="btn btn-primary" href="{{ route('roles.edit',$role->id) }}">Edit</a>
-           
+            @endcan
+
+            @can('role-delete')           
                 {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
                     {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                 {!! Form::close() !!}
+            @endcan
         </td>
     </tr>
     @endforeach
