+++ resources/views/users/index.blade.php
@@ -8,7 +8,9 @@
             <h2>Users Management</h2>
         </div>
         <div class="pull-right">
+          @can('user-create')
             <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
+          @endcan
         </div>
     </div>
 </div>
@@ -25,6 +27,7 @@
  <tr>
    <th>No</th>
    <th>Name</th>
+   <th>Department</th>
    <th>Email</th>
    <th>Roles</th>
    <th width="280px">Action</th>
@@ -33,6 +36,7 @@
   <tr>
     <td>{{ ++$i }}</td>
     <td>{{ $user->name }}</td>
+    <td>{{ $user->department->name }}</td>
     <td>{{ $user->email }}</td>
     <td>
       @if(!empty($user->getRoleNames()))
@@ -42,11 +46,17 @@
       @endif
     </td>
     <td>
-       <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
-       <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+      <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+
+      @can('user-edit')
+        <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+      @endcan
+
+      @can('user-delete')
         {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-            {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+        {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
         {!! Form::close() !!}
+      @endcan
     </td>
   </tr>
  @endforeach
